card-status-button moved inside app and card_status button inside student-list and studentcard component both is replaced by this card-status-button component
